Moved all collections into a seperate module.
