+++ cobbler/collections/collection_images.py
@@ -12,7 +12,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 """
 
 from actions import litesync
-from cobbler import collection
+from collections import collection
 from cobbler import item_image as image
 from cobbler import utils
 from cobbler.cexceptions import CX
